Move `ToolbarItem` inside `toolbar-item` subfolder
