+++ packages/components/src/toolbar/toolbar-dropdown-menu/index.js
@@ -6,7 +6,7 @@ import { forwardRef, useContext } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import ToolbarItem from '../../toolbar-item';
+import ToolbarItem from '../toolbar-item';
 import ToolbarContext from '../toolbar-context';
 import DropdownMenu from '../../dropdown-menu';
