+++ packages/components/src/toolbar/toolbar-group/toolbar-group-collapsed.js
@@ -8,7 +8,7 @@ import { useContext } from '@wordpress/element';
  */
 import DropdownMenu from '../../dropdown-menu';
 import ToolbarContext from '../toolbar-context';
-import ToolbarItem from '../../toolbar-item';
+import ToolbarItem from '../toolbar-item';
 
 function ToolbarGroupCollapsed( { controls = [], toggleProps, ...props } ) {
 	// It'll contain state if `ToolbarGroup` is being used within
